fix(compile): fix 66 compile errors - sqlx Option types, Record type mismatches, E0716 temporaries
- state.rs: remove conflicting FromRef<AppState> impl (E0119)
- main.rs: fix minijinja::Value::as_f64() -> as_i64() (E0599)
- db.rs: add .unwrap_or(0) for COUNT queries, .flatten() for fetch_optional (E0308)
- db.rs: fix temporary borrowed string in seed_system_settings (E0716)
- dashboard.rs: fix all COUNT(*) queries returning Option<i32> (E0308)
- api/customers.rs, api/system.rs: fix COUNT queries (E0308)
- contacts.rs: fix if/else Record type mismatch; extract bind params to let (E0308/E0716)
- customers.rs: fix if/else Record type mismatch; extract bind params to let (E0308/E0716)
- wan.rs: fix if/else Record type mismatch (E0308)
- invoices.rs: fix as_deref() on String; fix temp string borrow (E0599/E0716)
- auth.rs: cast row.id as i64 for create_session/log_login (E0308)
- users.rs: fix COUNT queries (E0308)
- recovery.rs: add .flatten() for fetch_optional (E0308)

+++ src/routes/auth.rs
@@ -70,7 +70,7 @@ pub async fn login(
                 // Create session
                 let token = match auth_utils::create_session(
                     &state.db,
-                    row.id,
+                    row.id as i64,
                     ip.as_deref(),
                     ua.as_deref(),
                     8,
@@ -93,7 +93,7 @@ pub async fn login(
                 .await
                 .ok();
 
-                audit::log_login(&state.db, row.id, &row.username, ip.as_deref(), true).await;
+                audit::log_login(&state.db, row.id as i64, &row.username, ip.as_deref(), true).await;
 
                 let redirect_to = form
                     .redirect_to
@@ -111,7 +111,7 @@ pub async fn login(
                     .unwrap()
             } else {
                 // Log failed attempt
-                audit::log_login(&state.db, row.id, &row.username, ip.as_deref(), false).await;
+                audit::log_login(&state.db, row.id as i64, &row.username, ip.as_deref(), false).await;
                 render_login_error(&state, "Ungültige Zugangsdaten")
             }
         }
